+++ docs/pages/changelog.md
@@ -2,6 +2,13 @@
 
 The following versions have been released:
 
+## 0.2.0-preview7
+
+* Upgrade to the final Blazor 3.2.0 WebAssembly version.
+* Added UserDataFolder to the BlazorNewEdgeWebView and BlazorWebView variants
+  for WPF to set the WebView2 userdata folder.
+  Fixes: https://github.com/jspuij/BlazorWebView/issues/15
+
 ## 0.2.0-preview6
 
 * Upgraded to the latest Blazor 3.2.0-preview5